Update .env and fixed texting logic
Initially I was checking for the email address "@dedicatedcoaching" but
apparently Courtney uses multiple email addresses and this will
occasionally result in me missing an email. To fix this, I am now
checking the Subject line for "Dedicated Coaching". I had to access the
full format of the emails by passing in {format: "full"} in the get
request for Subjects to start trickling through.